+++ src/Connection.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Blrf\Dbal;
 
 use React\Promise\PromiseInterface;
+use Blrf\Dbal\ResultStream;
 
 /**
  * Connection interface
@@ -31,6 +32,11 @@ interface Connection
      * @return PromiseInterface<Result>
      */
     public function execute(string $sql, array $params = []): PromiseInterface;
+
+    /**
+     * Execute query and return stream
+     */
+    public function stream(string $sql, array $params = []): ResultStream;
 
     /**
      * Get native connection