+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ## main-dev (2024-03-17)
 
 - QueryBuilder and Condition/ConditionGroup now have `fromArray()` and `toArray()` methods.
+
+## main-dev (2024-03-25)
+
+- Offical docs page: https://blrf.net/dbal/
+- Full coverage
+- Condtion/ConditionBuilder support for more operators (lg, lge, like, isNull, ...)